Use object-oriented Ruby to model a public library (w/ three classes: Library, Shelf, & Book). The library should be aware of a number of shelves. Each shelf should know what books it contains. Make the book object have "enshelf" and "unshelf" methods that control what shelf the book is sitting on. The library should have a method to report all books it contains.
Use object-oriented Ruby to model a public library (w/ three classes: Library, Shelf, & Book). The library should be aware of a number of shelves. Each shelf should know what books it contains. Make the book object have "enshelf" and "unshelf" methods that control what shelf the book is sitting on. The library should have a method to report all books it contains.

Note: this should not be a Rails app - just a single file with three classes (plus commands at the bottom showing it works) is all that is needed.